+++ src/Components/Event.js
@@ -1,3 +1,17 @@
-export default function Event() {
-  return;
+import Attendees from "./Attendees";
+
+export default function Event({ event, updateEventAttendance }) {
+  return (
+    <li className="event">
+      <img src={event.eventImage} alt={event.name} />
+      <h5>
+        {event.name} {event.eventType}
+      </h5>
+      <br />
+      <span>Organized by: {event.organizer} </span>
+      <br />
+
+      <Attendees event={event} updateEventAttendance={updateEventAttendance} />
+    </li>
+  );
 }
